feat: token inheritance (#29)
~~VERY draft~~ PR, the purpose of this small spike of relative success
is to consider if the Token related methods could be abstracted in a
general usage library. Ideally, a contract is a Token as a secondary
logic, that enables the transfer of amounts (private, commited --hiding
point--, and public).

The focus of this spike was to:
- move every `#[contract_library_method]` into a separate file that
declares `TokenLib`
- this was done by changing the expected `storage` into the actual
storage shape that's being expected on the library method (i.e.
`public_balances: Map<AztecAddress, ...>`). In that way, the `Storage`
struct doesn't need to be known in this part of the code
- some difficulties appeared when the library needs to call
`Token::at(this)`, but could be circumvented with raw
`call_public_function`, tho parsing the arguments into a `[Field;
ARGS_COUNT]` can be tricky
- move every method into a comptime function that injects them as plain
text
- `transfer_private_to_private` doesn't need to be declared inside
Contract codebase, it get's injected as plain text at compilation time
- surprisingly brillig code such as `balance_of_private` cannot be
injected in the comptime module (this raises some questions about
brillig --unconstrained-- code)
- sadly, imports need to be present in the Contract itself, and cannot
be passed by the module
- base dependencies can be reduced if passing more and more logic into
the TokenLib
- this base dependencies could play a good role in making flavours out
of libraries (the one that Contract inherits are the ones that end up
being used, i.e. encrypting logs)

The result is that a contract may inherit the `Token` functionalities,
as long as it has compatible storage (with `private_balances`,
`total_supply` and `public_balances`) with the correct shape, else it
fails to compile.

The dev may then inherit methods as `TokenLib::mint_private_amount` and
enjoy all of the contract's external calls at the cost of adding
`#[token]` tag (and compilation time).

Challenges:
- `Token:at(this)` lines needed to be commented but can be (a) replaced
for the hacky `call`, or (b) be defined at the time of compiling (i.e.
if the Contract is called `MyAmm` then do `MyAmm::at(this)`)
- working LSPless, once moved inside `#[token]`, the code is processed
as plain text until sent to compile, this creates an awful dev
experience where every error is pointed at the tag line
- Should Token code and TokenLib be dependant?
  - Token depends on TokenLib
- TokenLib needs to call sometimes `Token::at(this)` or expect some
methods as the recursive getter to exist
- Can we "not require Contract" to compile to be able to do
`IToken::at(this).recurse...`?
- Which methods are indispensable for TokenLib to work? and which others
are user entrypoints (ie. transfers)
- By placing all "emits" into the Token code (not the TokenLib), then
the main Contract may select the log emit strategy by importing a valid
`encode_and_encrypt_note` in the main contract declaration
- The partial note emit is tricky because is imported from the partial
note trait (which imports itself an encryption strategy iirc)